requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
psycopg2
psycopg2[binary]
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1
mvc-flask
gunicorn==21.2.0
marshmallow==3.20.1
redis==5.0.1
flask-cors
yfinance
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
pytest==7.4.0
pytest-flask==1.2.0
pytest-mock==3.11.1